refactor: move ClustersTable to feature
